requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
APScheduler==3.10.4
python-dateutil==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
Werkzeug==3.0.4
email-validator==2.2.0
reportlab==4.2.5
gunicorn==23.0.0
